Com_HashFileName(): replaced hashing logic with xxHash (https://github.com/Cyan4973/xxHash). In some cases textures/models were not loaded because their name hash matched with other texture/model, which already failed lo load.

+++ renderer/r_local.h
@@ -62,7 +62,7 @@ typedef enum
 typedef struct image_s
 {
 	char name[MAX_QPATH]; // Game path, including extension.
-	long hash; // Used to speed up searching. //mxd. Changed: stores hash of name without extension!
+	uint hash; // Used to speed up searching. //mxd. Changed: stores hash of name without extension!
 	imagetype_t type;
 
 	int width; // Source image size.
